Change homepage button to a component
